third_party/WebKit/LayoutTests/http/tests/htmlimports/csp-import-block-but-self-expected.txt

CONSOLE ERROR: line 2: Uncaught EvalError: Refused to evaluate a string as JavaScript because 'unsafe-eval' is not an allowed source of script in the following Content Security Policy directive: "script-src 'self'".

CONSOLE ERROR: Refused to load the script 'http://localhost:8000//htmlimports/resources/cors-js.cgi' because it violates the following Content Security Policy directive: "script-src 'self'".

CONSOLE ERROR: line 6: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'self'". Either the 'unsafe-inline' keyword, a hash ('sha256-wnLyUqsXhUuQ5NVNJO+JjNabqB1yo9OPLLJWPopauN0='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE MESSAGE: line 4: PASS
CONSOLE MESSAGE: line 4: PASS
CONSOLE MESSAGE: line 4: PASS
CONSOLE MESSAGE: line 4: PASS
CONSOLE MESSAGE: line 4: PASS
CONSOLE MESSAGE: line 4: PASS
